requirements.txt
fastapi
python-multipart
uvicorn[standard]
python-dotenv
requests
httpx
pydantic
reportlab
google-genai
